% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containers.R
\name{call_container}
\alias{call_container}
\title{Construct the code to call a container using a specific program}
\usage{
call_container(cmd, config_key, workdir)
}
\arguments{
\item{cmd}{command to execute in the container}

\item{config_key}{yml key in the config.yml file. The path to the container
will be constructed using config[[config_key]]$container}

\item{workdir}{workfolder to bind to the container}
}
\value{
a character vector of captured code
}
\description{
call_container uses three arguments to construct a call to a container:
}
\examples{
\dontrun{
call_container()
}
}
