% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstreamGWAS-package.R
\docType{package}
\name{downstreamGWAS-package}
\alias{downstreamGWAS}
\alias{downstreamGWAS-package}
\title{downstreamGWAS: Standardized downstream pipelines for GWAS}
\description{
Standardized analysis for genome-wide summary statistics.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://arvidharder.com/downstreamGWAS/}
}

}
\keyword{internal}
