% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{effective_n}
\alias{effective_n}
\title{Calculate effective N}
\usage{
effective_n(cases, controls)
}
\arguments{
\item{cases}{Number of cases}

\item{controls}{Number of controls}
}
\value{
a numeric value
}
\description{
Calculate effective N
}
\examples{
effective_n(1000, 3000)
}
