% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-analyze.R
\name{get_snp_ids}
\alias{get_snp_ids}
\title{Get the unique set of SNP, CHR, POS EffectAllele, OtherAllele, ID from a set of summary statistics}
\usage{
get_snp_ids(dset)
}
\arguments{
\item{dset}{a [arrow::open_dataset()] object}
}
\value{
a [dplyr::tibble()]
}
\description{
Get the unique set of SNP, CHR, POS EffectAllele, OtherAllele, ID from a set of summary statistics
}
\examples{
\dontrun{
dset <- arrow::open_dataset("path_to_dsets")
snpids <- get_snp_ids(dset)
}
}
