% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{get_system_paths}
\alias{get_system_paths}
\title{Read downstreamGWAS filepaths from yaml file
downstreamGWAS manages external filepaths through two files:}
\usage{
get_system_paths()
}
\value{
a nested list of filepaths
}
\description{
The param.yml file which contains filepaths to reference data and software containers
all filepaths start at reference/ or containers/ respectively. The param.yml file
is bundled with the downstreamGWAS package.
}
\details{
Secondly, the config.yml file which contains the local configuaration paramerers,
such as the path to the downstreamGWAS data folder, and how to call apptainer/singularity.
The config.yml file needs to be setup locally.
}
\examples{
\dontrun{
get_system_paths()
}
}
