% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldsc-r.R
\name{irwl_ldsc}
\alias{irwl_ldsc}
\title{Run LDscore regression --h2 flag.}
\usage{
irwl_ldsc(df, ld_size, step1_chi2 = 30, step2_chi2 = NULL, blocks = 200)
}
\arguments{
\item{df}{a data.frame containing three columns: c("Z", "N", "L2")
Z: Z score of association per snp
Z: sample size of variant
L2: LD score of variant}

\item{ld_size}{number of variants used to calculate LD score (Typically MAF > 0.05)}

\item{step1_chi2}{chi2 threshold for step 1 filter}

\item{step2_chi2}{chi2 threshold for step 2 filter: max(0.001 * max(ld_df$N), 80)}

\item{blocks}{number of blocks to use in the jackknife regression. NOT YET IMPLEMENTED}
}
\value{
a [dplyr::tibble()] with two columns: "int" and "obs_h2"
}
\description{
Run LDscore regression --h2 flag.
}
\examples{
\dontrun{
irwl_ldsc(df, ld_size = ld_size)
}
}
