% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldsc.R
\name{ldsc_rg}
\alias{ldsc_rg}
\title{Estimate genetic correlation between two GWAS using LDSC}
\usage{
ldsc_rg(parent_folder, parent_folder2, outdir, workdir = tempdir())
}
\arguments{
\item{parent_folder}{filepath to tidyGWAS output}

\item{parent_folder2}{filepath to tidyGWAS output}

\item{outdir}{Folder to save logfile in}

\item{workdir}{temporary working folder}
}
\value{
code to run in slurm
}
\description{
Estimate genetic correlation between two GWAS using LDSC
}
\examples{
\dontrun{
ldsc_rg("my_sumstats/tidygwas/height2022", "my_sumstats/tidygwas/height2022", "~/")
}
}
