% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ls_sumstats}
\alias{ls_sumstats}
\title{List the available sumstats}
\usage{
ls_sumstats(folder = NULL)
}
\arguments{
\item{folder}{can be used to specify where the summary statistics are.
Default value is NULL, and the sumstats_folder parameter will be used from the
config.yaml file}
}
\value{
a [dplyr::tibble()]
}
\description{
List the available sumstats
}
\examples{
\dontrun{
ls_sumstats()
}

}
