% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldsc.R
\name{parse_ldsc_h2}
\alias{parse_ldsc_h2}
\title{Parse the output of LDSC --h2}
\usage{
parse_ldsc_h2(path)
}
\arguments{
\item{path}{to log file}
}
\value{
a tibble
}
\description{
Parse the output of LDSC --h2
}
\examples{
\dontrun{
parse_ldsc_h2("ldsc_h2.log")
}
}
