% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldsc.R
\name{parse_ldsc_rg}
\alias{parse_ldsc_rg}
\title{Read in the output of LDSC --rg}
\usage{
parse_ldsc_rg(path)
}
\arguments{
\item{path}{path to LDSC rg output file}
}
\value{
a tibble
}
\description{
Read in the output of LDSC --rg
}
\examples{
\dontrun{
parse_ldsc_rg("path/to/ldsc_rg.results")
}

}
