% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{ranges_to_bed}
\alias{ranges_to_bed}
\title{Convert plink ranges file to bed}
\usage{
ranges_to_bed(path)
}
\arguments{
\item{path}{filepath to tidyGWAS folder}
}
\value{
bed file with clumps
}
\description{
Convert plink ranges file to bed
}
\examples{
\dontrun{
ranges_to_bed("/path/to/tidyGWAS")
}
}
