% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fileformats.R
\name{read_ldsc_h2}
\alias{read_ldsc_h2}
\title{read the output of LDSC --h2}
\usage{
read_ldsc_h2(path)
}
\arguments{
\item{path}{to log file}
}
\value{
a tibble
}
\description{
read the output of LDSC --h2
}
\examples{
\dontrun{
parse_ldsc_h2("ldsc_h2.log")
}
}
