% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clumping.R
\name{run_clumping}
\alias{run_clumping}
\title{Construct genetic loci from GWAS}
\usage{
run_clumping(parent_folder)
}
\arguments{
\item{parent_folder}{Folder to sumstats cleaned with tidyGWAS. see [tidyGWAS::tidyGWAS()] output_folder}
}
\value{
a filepath to a clumping script
}
\description{
Construct genetic loci from GWAS
}
\examples{
\dontrun{
run_clumping("/my_sumstat/cleaned/")
}
}
