% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc.R
\name{run_coloc}
\alias{run_coloc}
\title{Run coloc::coloc.abf on tidyGWAS data}
\usage{
run_coloc(
  parent_dir,
  parent_dir2,
  chr,
  start,
  end,
  min_pval = 5e-08,
  trait_type1 = c("guess", "cc", "quant"),
  trait_type2 = c("guess", "cc", "quant"),
  p1 = 1e-04,
  p2 = 1e-04,
  p12 = 1e-05
)
}
\arguments{
\item{parent_dir}{tidyGWAS directory of GWAS}

\item{parent_dir2}{tidyGWAS directory of GWAS}

\item{chr}{chromosome}

\item{start}{start of region}

\item{end}{end of region}

\item{min_pval}{minimum pval required to proceed with coloc in trait2}

\item{trait_type1}{quantitative or case-control?}

\item{trait_type2}{quantitative or case-control?}

\item{p1}{prior for colof.abf}

\item{p2}{prior for colof.abf}

\item{p12}{prior for colof.abf}
}
\value{
output of \code{\link[coloc:coloc.abf]{coloc::coloc.abf()}}
}
\description{
Run coloc::coloc.abf on tidyGWAS data
}
\examples{
\dontrun{
run_coloc("path/trait1", "path/trait2")
}
}
