% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldsc.R
\name{run_ldsc}
\alias{run_ldsc}
\title{Run munge LDSC and LDSC -h2 from tidyGWAS}
\usage{
run_ldsc(parent_folder)
}
\arguments{
\item{parent_folder}{Folder to sumstats cleaned with tidyGWAS. see [tidyGWAS::tidyGWAS()] output_folder}
}
\value{
a path to slurm script
}
\description{
Run munge LDSC and LDSC -h2 from tidyGWAS
}
\examples{
\dontrun{
script_location <- run_ldsc("my_sumstats/tidygwas/height2022")
}
}
