% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gctb.R
\name{run_mbat_combo}
\alias{run_mbat_combo}
\title{run mBAT-combo gene-test in GCTB}
\usage{
run_mbat_combo(
  parent_folder,
  ...,
  write_script = TRUE,
  outfolder = NULL,
  thread_num = 10
)
}
\arguments{
\item{parent_folder}{filepath to a tidyGWAS folder}

\item{...}{arguments to slurm}

\item{write_script}{should the code be written to disk?}

\item{outfolder}{Where to write the output}

\item{thread_num}{number of threads to use}
}
\value{
a character vector with the script or a filepath
}
\description{
run mBAT-combo gene-test in GCTB
}
\examples{
\dontrun{
mbat_combo("path_to_tidyGWAS/folder/sumstat1")
}
}
