% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbayesrc.R
\name{run_sbayesrc}
\alias{run_sbayesrc}
\title{Create a slurmjob for SbayesRC}
\usage{
run_sbayesrc(path)
}
\arguments{
\item{path}{filepath to tidyGWAS_hivestyle folder}
}
\value{
a slurm script
}
\description{
Create a slurmjob for SbayesRC
}
\examples{
\dontrun{
run_sbayesrc("path/to/tidyGWAS_hivestyle")
}
}
