% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gctb.R
\name{run_sbayess}
\alias{run_sbayess}
\title{Run Sbayes-S with tidyGWAS structure}
\usage{
run_sbayess(parent_folder, ..., write_script = c("yes", "no"))
}
\arguments{
\item{parent_folder}{filepath to a [tidyGWAS] folder}

\item{...}{pass arguments to [slurm_header()]}

\item{write_script}{should the captured code be written to disk in a .sh file?}
}
\value{
a filepath or character vector
}
\description{
Run Sbayes-S with tidyGWAS structure
}
\examples{
\dontrun{
run_sbayess()
}
}
