% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gctb.R
\name{run_sbayess_req}
\alias{run_sbayess_req}
\title{Check if required files exist for Sbayes-S}
\usage{
run_sbayess_req()
}
\description{
Check if required files exist for Sbayes-S
}
\examples{
\dontrun{
run_sbayess_req()
}
}
