% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldsc.R
\name{run_sldsc}
\alias{run_sldsc}
\title{run Stratified LDscore regression on tidyGWAS formatted sumstats}
\usage{
run_sldsc(
  parent_folder,
  ldscore = c("superclusters", "clusters"),
  write_script = c("no", "yes")
)
}
\arguments{
\item{parent_folder}{Folder to sumstats cleaned with tidyGWAS. see [tidyGWAS::tidyGWAS()] output_folder}

\item{ldscore}{which cell-type atlas to use. Has to match an entry in the reference file}

\item{write_script}{Should the code be written to a bash script?}
}
\description{
run Stratified LDscore regression on tidyGWAS formatted sumstats
}
\examples{
\dontrun{
run_sldsc("path_to_tidyGWAS")
}
}
