% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldsc.R
\name{run_sldsc_cts}
\alias{run_sldsc_cts}
\title{Run stratified LDscore regression using the --h2-cts flag}
\usage{
run_sldsc_cts(parent_folder, cts_file, write_script = TRUE, ..., out = NULL)
}
\arguments{
\item{parent_folder}{the filepath to a [tidyGWAS::tidyGWAS()] folder}

\item{cts_file}{filepath to cts file}

\item{write_script}{should the code be written to a file?}

\item{...}{optional slurm argument}

\item{out}{filepath to output directory. Defaults to "sldsc" in the ldsc folder}
}
\value{
a character vector
}
\description{
Run stratified LDscore regression using the --h2-cts flag
}
\examples{
\dontrun{
  run_sldsc_cts(tempdir(), "siletti2023.cts")
}

}
