% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbayesrc.R
\name{sbayesrc}
\alias{sbayesrc}
\title{Run SbayesRC from R with containers}
\usage{
sbayesrc(workdir, ldm, ma_file, annot, out, thread_imp = 4, thread_rc = 4)
}
\arguments{
\item{workdir}{where will the files be written to?}

\item{ldm}{path to ldmatrix}

\item{ma_file}{path to ma file}

\item{annot}{path to annotation file}

\item{out}{output folder}

\item{thread_imp}{how many threads for imputation?}

\item{thread_rc}{how many threads for rescaling?}
}
\value{
character vector
}
\description{
Run SbayesRC from R with containers
}
\examples{
\dontrun{
sbayesrc(tempdir(), "path_to_ldm", "sumstats.ma", "annot.txt", out = tempdir())
}
}
