% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbayesrc.R
\name{sbayesrc_pipeline}
\alias{sbayesrc_pipeline}
\title{Run SbayesRC}
\usage{
sbayesrc_pipeline(paths)
}
\arguments{
\item{paths}{crated by `tidyGWAS_paths()`}
}
\description{
Run SbayesRC
}
\examples{
\dontrun{
}
}
