% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{schedule_slurm}
\alias{schedule_slurm}
\title{Construct a SLURM schedule object}
\usage{
schedule_slurm(
  time = "24:00:00",
  mem = "8gb",
  cpus_per_task = NULL,
  account = NULL,
  partition = NULL
)
}
\arguments{
\item{time}{SLURM time limit.}

\item{mem}{SLURM memory request.}

\item{cpus_per_task}{SLURM CPUs per task.}

\item{account}{Optional SLURM account.}

\item{partition}{Optional SLURM partition.}
}
\value{
A schedule object that can be passed to \verb{pipeline_*()} functions.
}
\description{
Construct a SLURM schedule object
}
