% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{Setup required filepaths for downstreamGWAS}
\usage{
setup(dir)
}
\arguments{
\item{dir}{in what directory should the config file be stored?}
}
\description{
Setup required filepaths for downstreamGWAS
}
