% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{setup_dsg}
\alias{setup_dsg}
\title{Setup downstreamGWAS config}
\usage{
setup_dsg(
  storage_root,
  sumstats_folder = NULL,
  container_dependency = "",
  force = FALSE
)
}
\arguments{
\item{storage_root}{Long-term storage root for references and containers.}

\item{sumstats_folder}{Optional default folder containing tidyGWAS datasets.}

\item{container_dependency}{Optional shell command to load runtime dependencies
on HPC (e.g. \code{"ml apptainer"}).}

\item{force}{Overwrite existing config file?}
}
\value{
Path to written config file.
}
\description{
Setup downstreamGWAS config
}
