% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm.R
\name{slurm_header}
\alias{slurm_header}
\title{Construct a slurm header}
\usage{
slurm_header(
  time = "30:00:00",
  mem = "8gb",
  output = NULL,
  account = NULL,
  partition = NULL,
  cpus_per_task = NULL
)
}
\arguments{
\item{time}{maximum run time for the job}

\item{mem}{amount of memory to allocate}

\item{output}{filepath to slurm output file}

\item{account}{slurm account to debit cost}

\item{partition}{slurm partition to use}

\item{cpus_per_task}{number of cpus per task}
}
\value{
a character vector
}
\description{
Construct a slurm header
}
\examples{
\dontrun{
slurm_header(time = "30:00:00", mem = "8gb", output = "output.txt",
account = "myaccount", partition = "my_partition", cpus_per_task = 10)
}
}
