% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{tidyGWAS_paths}
\alias{tidyGWAS_paths}
\title{Create the folder structure and filepaths for downstreamGWAS directory}
\usage{
tidyGWAS_paths(dir)
}
\arguments{
\item{dir}{filepath to folder}
}
\value{
a list of filepaths
}
\description{
Create the folder structure and filepaths for downstreamGWAS directory
}
\examples{
\dontrun{
 tidyGWAS_paths("gwas/height2022")
}
}
