% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{to_clumping}
\alias{to_clumping}
\title{Prepare tidyGWAS sumstats for PLINK clumping}
\usage{
to_clumping(hivestyle_path, output_dir)
}
\arguments{
\item{hivestyle_path}{path to tidyGWAS hivestyle dataset}

\item{output_dir}{directory to write sumstats.tsv}
}
\description{
Prepare tidyGWAS sumstats for PLINK clumping
}
\examples{
\dontrun{
to_clumping("/path/to/hivestyle/dataset", "/path/to/output/dir")
}
}
