% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{to_ldsc}
\alias{to_ldsc}
\title{Export tidyGWAS format to LDSC format}
\usage{
to_ldsc(parent_folder)
}
\arguments{
\item{parent_folder}{filepath to tidyGWAS folder}
}
\description{
Export tidyGWAS format to LDSC format
}
\examples{
\dontrun{
to_ldsc("my_sumstats/tidygwas/height2022")
}
}
