% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbayesrc.R
\name{wrapper_sbayesrc}
\alias{wrapper_sbayesrc}
\title{Capture code to run sbayesRC}
\usage{
wrapper_sbayesrc(paths, thread = 4)
}
\arguments{
\item{paths}{a list of filepaths, see [sbayesrc()] for which filepaths are required}

\item{thread}{number of threads to use}
}
\value{
a character vector
}
\description{
Capture code to run sbayesRC
}
\examples{
\dontrun{
wrapper_sbayesrc()
}
}
