% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{check_setup}
\alias{check_setup}
\title{Check that the configuration file has been correctly set up}
\usage{
check_setup()
}
\value{
text to terminal
}
\description{
Check that the configuration file has been correctly set up
}
\examples{
\dontrun{
check_setup()
}
}
