% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstreamGWAS-package.R
\docType{package}
\name{downstreamGWAS-package}
\alias{downstreamGWAS}
\alias{downstreamGWAS-package}
\title{downstreamGWAS: Standardized downstream pipelines for GWAS}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
