% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc.R
\name{format_coloc}
\alias{format_coloc}
\title{Format the output of \code{\link[coloc:coloc.abf]{coloc::coloc.abf()}}}
\usage{
format_coloc(coloc_obj, name)
}
\arguments{
\item{coloc_obj}{output of coloc call}

\item{name}{name of trait}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
Format the output of \code{\link[coloc:coloc.abf]{coloc::coloc.abf()}}
}
\examples{
\dontrun{
format_coloc(ob, "test-run")
}
}
