% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepath_structure.R
\name{get_system_paths}
\alias{get_system_paths}
\title{Read downstreamGWAS filepaths from yaml file}
\usage{
get_system_paths()
}
\value{
a nested list of filepaths
}
\description{
Read downstreamGWAS filepaths from yaml file
}
\examples{
\dontrun{
get_system_paths()
}
}
