% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-analysis.R
\name{meta_analyze_by_chrom}
\alias{meta_analyze_by_chrom}
\title{meta_analyze summary statistics, one chromosome at a time!}
\usage{
meta_analyze_by_chrom(dset, chrom)
}
\arguments{
\item{dset}{an object created by [arrow::open_dataset()]}

\item{chrom}{chromosome to meta-analyse}
}
\value{
a [dplyr::tibble()]
}
\description{
meta_analyze summary statistics, one chromosome at a time!
}
\examples{
\dontrun{
meta_analyze_by_crom(dset, chrom = "22")
}
}
