% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-sample-mr.R
\name{mr_on_tidyGWAS}
\alias{mr_on_tidyGWAS}
\title{Run two-sample mendelian randomisation using the TwoSampleMR package}
\usage{
mr_on_tidyGWAS(
  exposure_dir,
  outcome_dir,
  exposure_bed = NULL,
  bidirectional = FALSE
)
}
\arguments{
\item{exposure_dir}{path to tidyGWAS directory of the exposure}

\item{outcome_dir}{path to tidyGWAS directory of the outcome}

\item{exposure_bed}{Use a custom bed file to define lead SNPs? Default is NULL,
and downstreamGWAS will run \code{\link[=run_clumping]{run_clumping()}} if no bed file exists.}

\item{bidirectional}{run with outcome as exposure and exposure as outcome as well?}
}
\value{
a list
}
\description{
Run two-sample mendelian randomisation using the TwoSampleMR package
}
\examples{
\dontrun{
mr_on_tidyGWAS("exp_dir/trait1", "outcomes/trait2")
}
}
