% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{ranges_to_bed}
\alias{ranges_to_bed}
\title{Convert clumped.ranges output from Plink to bed format}
\usage{
ranges_to_bed(infile, out)
}
\arguments{
\item{infile}{*.clumped.ranges file}

\item{out}{filename of bed file that is created}
}
\value{
a bedfile
}
\description{
Convert clumped.ranges output from Plink to bed format
}
\examples{
\dontrun{
ranges_to_bed("/my_sumstat/cleaned/clumps.clumped.ranges", "/my_sumstat/cleaned/clumps.bed")
}
}
