% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{ranges_to_bed}
\alias{ranges_to_bed}
\title{Convert the output of plink clumping ranges to a flat file}
\usage{
ranges_to_bed(clump_dir)
}
\arguments{
\item{clump_dir}{directory where plink clumping output is stored}
}
\description{
Convert the output of plink clumping ranges to a flat file
}
\examples{
\dontrun{
ranges_to_bed("/path/to/clumping/dir")
}
}
