% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clumping.R
\name{run_clumping}
\alias{run_clumping}
\title{Construct genetic loci from GWAS}
\usage{
run_clumping(parent_folder, write_script = c("no", "yes"))
}
\arguments{
\item{parent_folder}{Folder to sumstats cleaned with tidyGWAS. see [tidyGWAS::tidyGWAS()] output_folder}

\item{write_script}{should a .sh file be written to the workdir with the code?}
}
\value{
a filepath to a clumping script
}
\description{
Construct genetic loci from GWAS
}
\examples{
\dontrun{
run_clumping("/my_sumstat/cleaned/")
}
}
