% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{run_clumping}
\alias{run_clumping}
\title{Run a clumping pipeline on a tidyGWAS sumstats}
\usage{
run_clumping(path)
}
\arguments{
\item{path}{filepath to tidyGWAS folder}
}
\value{
bed file with clumps
}
\description{
Run a clumping pipeline on a tidyGWAS sumstats
}
\examples{
\dontrun{
ranges_to_bed("/path/to/tidyGWAS")
}
}
