% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldsc.R
\name{run_ldsc}
\alias{run_ldsc}
\title{Run munge LDSC and LDSC -h2 from tidyGWAS}
\usage{
run_ldsc(parent_folder, ..., write_script = TRUE, use_effective_n = TRUE)
}
\arguments{
\item{parent_folder}{Folder to sumstats cleaned with tidyGWAS. see \code{\link[tidyGWAS:tidyGWAS]{tidyGWAS::tidyGWAS()}} output_folder}

\item{...}{Pass arguments to \code{\link[=slurm_header]{slurm_header()}}}

\item{write_script}{Should the code be written to a bash script?}

\item{use_effective_n}{Should sample size be computed with the effective_n formula if
CaseN and ControlN is present?}
}
\value{
a path to slurm script
}
\description{
Run munge LDSC and LDSC -h2 from tidyGWAS
}
\examples{
\dontrun{
script_location <- run_ldsc("my_sumstats/tidygwas/height2022")
}
}
