% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gctb.R
\name{run_sbayesrc}
\alias{run_sbayesrc}
\title{Run sbayerc with tidyGWAS structure}
\usage{
run_sbayesrc(
  parent_folder,
  ...,
  write_script = TRUE,
  thread_rc = 8,
  thread_imp = 4,
  use_effective_n = FALSE,
  repair_EAF = NULL
)
}
\arguments{
\item{parent_folder}{path to tidyGWAS folder}

\item{...}{pass arguments to \code{\link[=slurm_header]{slurm_header()}}}

\item{write_script}{Should the script be written to a file on disk?}

\item{thread_rc}{threads for rescaling}

\item{thread_imp}{threads for imputing}

\item{use_effective_n}{Should an attempt be made to calculate effective N}

\item{repair_EAF}{Should EAF be repaired? If so, provide a path to a file with columns RSID, EffectAllele, OtherAllele, EAF}
}
\value{
a filepath or character vector
}
\description{
Run sbayerc with tidyGWAS structure
}
\examples{
\dontrun{
run_sbayesrc()
}
}
