% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbayesrc.R
\name{run_sbayesrc}
\alias{run_sbayesrc}
\title{Run sbayerc with tidyGWAS structure}
\usage{
run_sbayesrc(parent_folder, write_script = c("no", "yes"), ...)
}
\arguments{
\item{parent_folder}{Folder to sumstats cleaned with tidyGWAS. see [tidyGWAS::tidyGWAS()] output_folder}

\item{write_script}{Should the code be written to a bash script?}

\item{...}{
  Arguments passed on to \code{\link[=sbayesrc]{sbayesrc}}
  \describe{
    \item{\code{workdir}}{work directory}
    \item{\code{ldm}}{filepath to ldm folder}
    \item{\code{ma_file}}{filepath to.am file}
    \item{\code{imp_file}}{filepath to imputed .ma file}
    \item{\code{annot}}{filepath to the annotation file}
    \item{\code{out}}{filepath prefix to outfiles}
    \item{\code{thread}}{number of threads}
  }}
}
\value{
a filepath or character vector
}
\description{
Run sbayerc with tidyGWAS structure
}
\examples{
\dontrun{
run_sbayesrc()
}
}
