% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gctb.R
\name{run_sbayesrc}
\alias{run_sbayesrc}
\title{Run sbayerc with tidyGWAS structure}
\usage{
run_sbayesrc(
  parent_folder,
  ...,
  write_script = c("no", "yes"),
  thread_rc = 8,
  thread_imp = 4
)
}
\arguments{
\item{parent_folder}{Folder to sumstats cleaned with tidyGWAS. see \code{\link[tidyGWAS:tidyGWAS]{tidyGWAS::tidyGWAS()}} output_folder}

\item{...}{
  Arguments passed on to \code{\link[=sbayesrc]{sbayesrc}}
  \describe{
    \item{\code{workdir}}{where will the files be written to?}
    \item{\code{ldm}}{path to ldmatrix}
    \item{\code{ma_file}}{path to ma file}
    \item{\code{annot}}{path to annotation file}
    \item{\code{out}}{output folder}
  }}

\item{write_script}{Should the code be written to a bash script?}

\item{thread_rc}{threads for rescaling}

\item{thread_imp}{threads for imputing}
}
\value{
a filepath or character vector
}
\description{
Run sbayerc with tidyGWAS structure
}
\examples{
\dontrun{
run_sbayesrc()
}
}
