% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gctb.R
\name{run_sbayesrc}
\alias{run_sbayesrc}
\title{Run sbayerc with tidyGWAS structure}
\usage{
run_sbayesrc(
  parent_folder,
  ...,
  write_script = TRUE,
  thread_rc = 8,
  thread_imp = 4
)
}
\arguments{
\item{parent_folder}{Folder to sumstats cleaned with tidyGWAS. see \code{\link[tidyGWAS:tidyGWAS]{tidyGWAS::tidyGWAS()}} output_folder}

\item{...}{Pass arguments to \code{\link[=slurm_header]{slurm_header()}}}

\item{write_script}{Should the code be written to a bash script?}

\item{thread_rc}{threads for rescaling}

\item{thread_imp}{threads for imputing}
}
\value{
a filepath or character vector
}
\description{
Run sbayerc with tidyGWAS structure
}
\examples{
\dontrun{
run_sbayesrc()
}
}
