% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gctb.R
\name{run_sbayess}
\alias{run_sbayess}
\title{Run Sbayes-S with tidyGWAS structure}
\usage{
run_sbayess(
  parent_folder,
  ...,
  write_script = TRUE,
  pi = "0.01",
  hsq = "0.5",
  num_chains = "4",
  chain_length = "25000",
  burn_in = "5000",
  seed = "2023",
  thread = "8"
)
}
\arguments{
\item{parent_folder}{filepath to a [tidyGWAS] folder}

\item{...}{pass arguments to [slurm_header()]}

\item{write_script}{should the captured code be written to disk in a .sh file?}

\item{pi}{argument passed to sbayes}

\item{hsq}{argument passed to sbayes}

\item{num_chains}{argument passed to sbayes}

\item{chain_length}{argument passed to sbayes}

\item{burn_in}{argument passed to sbayes}

\item{seed}{argument passed to sbayes}

\item{thread}{argument passed to sbayes}
}
\value{
a filepath or character vector
}
\description{
Run Sbayes-S with tidyGWAS structure
}
\examples{
\dontrun{
run_sbayess()
}
}
