% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbayesrc.R
\name{sbayesrc}
\alias{sbayesrc}
\title{Capture code to run sbayesRC}
\usage{
sbayesrc(workdir, ldm, ma_file, imp_file, annot, out, thread = 4)
}
\arguments{
\item{workdir}{work directory}

\item{ldm}{filepath to ldm folder}

\item{ma_file}{filepath to.am file}

\item{imp_file}{filepath to imputed .ma file}

\item{annot}{filepath to the annotation file}

\item{out}{filepath prefix to outfiles}

\item{thread}{number of threads}
}
\value{
a character vector
}
\description{
Capture code to run sbayesRC
}
\examples{
\dontrun{
sbayesrc()
}
}
