% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{setup}
\alias{setup}
\title{Setup required filepaths for downstreamGWAS}
\usage{
setup(
  downstreamGWAS_folder,
  container_dependency = "",
  container_software = c("apptainer", "singularity")
)
}
\arguments{
\item{downstreamGWAS_folder}{Local filepath to where the downstreamGWAS data is stored}

\item{container_dependency}{Do you need to load singularity/apptainer on your HPC?
For example: "ml apptainer"}

\item{container_software}{Which container software do you use? "apptainer" or "singularity"}
}
\description{
Setup required filepaths for downstreamGWAS
}
\examples{
\dontrun{
setup()
}
}
