% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm.R
\name{slurm_header}
\alias{slurm_header}
\title{Construct a slurm header}
\usage{
slurm_header(
  time = "24:00:00",
  mem = "8gb",
  output = NULL,
  account = NULL,
  partition = NULL,
  cpus_per_task = NULL
)
}
\arguments{
\item{time}{time allocated to job}

\item{mem}{memory allocated to job (remember to use 'gb' ending)}

\item{output}{filepath to output slurm log file}

\item{account}{account}

\item{partition}{partition}

\item{cpus_per_task}{cpus per task}
}
\value{
A character vector with the slurm header
}
\description{
Construct a slurm header
}
\examples{
slurm_header()
}
