% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{to_clumping}
\alias{to_clumping}
\title{Convert tidyGWAS to file with RSID and P for clumping}
\usage{
to_clumping(path)
}
\arguments{
\item{path}{filepath to tidyGWAS folder}
}
\value{
writes out a tsv.gz file
}
\description{
Convert tidyGWAS to file with RSID and P for clumping
}
\examples{
\dontrun{
to_clumping("/path/to/tidyGWAS")
}
}
