% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldsc.R
\name{to_ldsc}
\alias{to_ldsc}
\title{Convert parquet hivestyle data to csv for LD-score regression}
\usage{
to_ldsc(dir)
}
\arguments{
\item{dir}{filepath to the directory parent directory of tidyGWAS_hivestyle}
}
\value{
commandline code
}
\description{
Convert parquet hivestyle data to csv for LD-score regression
}
\examples{
\dontrun{
code <- to_ldsc("files/sumstats/height2022")
}
}
