% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{to_ldsc}
\alias{to_ldsc}
\title{Export tidyGWAS format to LDSC format}
\usage{
to_ldsc(parent_folder, use_effective_n = TRUE)
}
\arguments{
\item{parent_folder}{filepath to tidyGWAS folder}

\item{sample_size}{Should an attempt be made to calculate effective N and use that as N?}
}
\description{
Export tidyGWAS format to LDSC format
}
\examples{
\dontrun{
to_ldsc("my_sumstats/tidygwas/height2022")
}
}
