% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbayesrc.R
\name{to_ma}
\alias{to_ma}
\title{Transform tidyGWAS to COJO .ma format}
\usage{
to_ma(paths)
}
\arguments{
\item{paths}{crated by `tidyGWAS_paths()`}
}
\value{
a list of filepaths
}
\description{
Transform tidyGWAS to COJO .ma format
}
\examples{
\dontrun{
  tidyGWAS_paths("/home/arvhar/tidyGWAS_cleaned/height2022")
}

}
