% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{to_ma}
\alias{to_ma}
\title{Convert tidyGWAS to COJO .ma format}
\usage{
to_ma(parent_folder, out)
}
\arguments{
\item{parent_folder}{filepath to tidyGWAS folder}

\item{out}{output for .ma file default is the tidyGWAS folder structure}
}
\description{
Convert tidyGWAS to COJO .ma format
}
\examples{
\dontrun{
to_ma("/path/tidyGWAS_sumstats/a_sumstat")
}
}
