% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clumping.R
\name{to_plink_clumping}
\alias{to_plink_clumping}
\title{Convert tidyGWAS hivestyle partitioning to compatible format for plink clumping}
\usage{
to_plink_clumping(dir)
}
\arguments{
\item{dir}{filepath to tidyGWAS folder}
}
\value{
writes a file to disk
}
\description{
Convert tidyGWAS hivestyle partitioning to compatible format for plink clumping
}
\examples{
\dontrun{
paths <- tidyGWAS_paths("/my_sumstat/cleaned/")
to_plink_clumping(paths)
}
}
