% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containers.R
\name{with_container}
\alias{with_container}
\title{Run a command in a container}
\usage{
with_container(exe_path, code, config_key, workdir)
}
\arguments{
\item{exe_path}{filepath inside the container to software to run}

\item{code}{code to run in the container in the software specified by exe_path}

\item{config_key}{yml key in the config.yml file. The path to the container}

\item{workdir}{workfolder to bind to the container}
}
\value{
a character vector of captured code
}
\description{
Run a command in a container
}
\examples{
\dontrun{
with_container()
}
}
