% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containers.R
\name{with_container}
\alias{with_container}
\title{Run arbitrary code inside a container}
\usage{
with_container(
  code,
  image,
  workdir,
  env = NULL,
  setup_exists = FALSE,
  R_code = FALSE
)
}
\arguments{
\item{code}{code to be executed inside a container.}

\item{image}{which container should be used? Used to index into params.yml}

\item{workdir}{filepath to a directory. Will be used as the working directory inside the container}

\item{env}{pass environmental variables to the container, in format:
"KEY=VALUE"
Currently only supports passing one variable}

\item{setup_exists}{logical. If TRUE, the workdir, reference_dir and container_dependecy paths are assumed to exist.
bind paths and code to load apptainer/singularity will not be written out to the script}

\item{R_code}{Running R Code using the format: R -e "$code" is challening due to escaping of
quotes and special characters. If TRUE, the code will be run using R -e "$code"}
}
\value{
a character vector of captured code
}
\description{
Run arbitrary code inside a container
}
\examples{

with_container(
 code = "echo hello",
 image = "R",
 workdir = tempdir()
 )

}
