% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gctb.R
\name{wrapper_sbayesrc}
\alias{wrapper_sbayesrc}
\title{Capture code to run sbayesRC}
\usage{
wrapper_sbayesrc(paths, thread_imp = 4, thread_rc = 8)
}
\arguments{
\item{paths}{a list of filepaths, see \code{\link[=sbayesrc]{sbayesrc()}} for which filepaths are required}

\item{thread_imp}{number of threads to for the imputation step}

\item{thread_rc}{number of threads for the rescaling step}
}
\value{
a character vector
}
\description{
Capture code to run sbayesRC
}
\examples{
\dontrun{
wrapper_sbayesrc()
}
}
